# Conflicts:
#	contracts/fdc/implementation/FdcVerification.sol
#	contracts/userInterfaces/IFdcVerification.sol

+++ contracts/userInterfaces/IFlareSystemsCalculator.sol
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: MIT
 pragma solidity >=0.7.6 <0.9;
 
-
 /**
  * FlareSystemsCalculator interface.
  */
@@ -10,7 +9,7 @@ interface IFlareSystemsCalculator {
     /// Event emitted when the registration weight of a voter is calculated.
     event VoterRegistrationInfo(
         address indexed voter,
-        uint24 indexed rewardEpochId,
+        uint32 indexed rewardEpochId,
         address delegationAddress,
         uint16 delegationFeeBIPS,
         uint256 wNatWeight,
